Trading/backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
MetaTrader5>=5.0.45,<6.0.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<3.0.0
ta>=0.11.0,<1.0.0
apscheduler>=3.10.0,<4.0.0
httpx>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
alembic>=1.13.0,<2.0.0
